// src/app/Models/Buttons.ts
export interface ITextButtons {
  topButton: 'Начать разговор' | 'Запись...' | 'Остановить воспроизведение';
  bottomButton:
    | 'Окончить разговор'
    | 'Воспроизвести запись'
    | 'Нет записей'
    | 'Остановить воспроизведение';
}